% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{configureAWS}
\alias{configureAWS}
\title{Configure AWS}
\usage{
configureAWS(keycsv, profile = "default", region = "us-east-2",
  output = "json")
}
\arguments{
\item{keycsv}{character path to csv file obtained from AWS containing the id and key.}

\item{profile}{character username to setup profile.}

\item{region}{character amazon region to use.}

\item{output}{character amazon default output from aws command line interface tool.}
}
\description{
Configures AWS using access key ID and secret access key provided in csv by user.
Need to run on master. This sets the default access key and id.
to get access key, click on name in top right of the management page, click security credentials, click access keys, click create access key.
}
