% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{dbxcli}
\alias{dbxcli}
\title{Interface to command line tool for dropbox}
\usage{
dbxcli(cmd, ..., ncores = 1)
}
\arguments{
\item{cmd}{String indicating command to run.}

\item{...}{arguments passed to command.}

\item{ncores}{number of vectors of commands.}
}
\description{
Interface to command line tool for dropbox
}
