% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadABIDE.R
\name{downloadABIDE}
\alias{downloadABIDE}
\title{Download ABIDE data for analysis.}
\usage{
downloadABIDE(outdir, force = FALSE)
}
\arguments{
\item{outdir}{A directory to save the output files.}

\item{force}{Always re-download existing files.}
}
\value{
Returns a data frame of ABIDE demographic data.
}
\description{
Download ABIDE data for analysis.
}
