% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadABIDE.R
\name{downloadABIDEFile}
\alias{downloadABIDEFile}
\title{Download ABIDE data for analysis.}
\usage{
downloadABIDEFile(url, destfile, force)
}
\arguments{
\item{url}{Undocumented arguments from Jeremy.}

\item{destfile}{Undocumented arguments from Jeremy.}

\item{force}{Undocumented arguments from Jeremy.}
}
\value{
Returns a file?
}
\description{
Download ABIDE data for analysis.
}
\examples{

# downloadABIDE('./', derivatives='falff')
}
