% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSimData.R
\name{genSimData}
\alias{genSimData}
\title{Generates data for simulation}
\usage{
genSimData(
  files,
  outfiles = NULL,
  n4sim,
  nMeas = NULL,
  mask = NULL,
  method = c("bootstrap", "synthetic"),
  lambda = 0.5
)
}
\arguments{
\item{files}{a vector of .nii or .nii.gz images.}

\item{outfiles}{a vector of images to save the output.}

\item{n4sim}{total number of measurements that we're simulating}

\item{nMeas}{Approximate average number of measurements per subject.}

\item{mask}{image mask where data exist.}

\item{method}{Method to generate data, either "synthetic" (i.e. multivariate normal) or "bootstrap."}

\item{lambda}{Simulated data are a convex coombination of normally distributed data with bootstrapped data. lambda=1 is fully bootstrapped data, lambda=0 is fully synthetic.}
}
\description{
This function loads images specified in filelist and adds signal to the
 images. The amount and structure of signal is determined by betaimg and X.
 The output images are equal the images specified by filelist plus betaimg
 times the column of X that is not in Xred. This column is first
 residualized to Xred.
}
\keyword{null}
\keyword{power}
\keyword{simulation}
