% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{memoryUse}
\alias{memoryUse}
\title{List memory usage of all objects}
\usage{
memoryUse(units = "MiB")
}
\arguments{
\item{units}{character what units to use, passed to format for object.size.}
}
\description{
List memory usage of all objects
}
