% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{parameterImage}
\alias{parameterImage}
\title{Creates parameter image in random locations for simulations}
\usage{
parameterImage(mask, parameterImage, rs, betas)
}
\arguments{
\item{mask}{String or niftiImage object indication the area to select sphere locations from.}

\item{parameterImage}{output nifti for the parameter image.}

\item{rs}{vector of radii for signal spheres.}

\item{betas}{vector of parameters for signal spheres. Signal is constant throughout the sphere.
Model parameter values are set to params = betas * sd(y)/sd(x).}
}
\value{
Returns the parameter image after writing it to file.path(outdir, 'signal.nii.gz').
}
\description{
Uses the mask variable to create length(rs) spheres of size rs with values equal to betas within the mask.
A random voxel is chosen within the mask and a sphere is placed at that location.
The spheres are masked by the mask image so that no parameter values exist outside the mask.
}
