% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{simSetup}
\alias{simSetup}
\title{Sets up a Neuroimaging bootstrap-based simulation given a set of images and covariates.}
\usage{
simSetup(images, data, outdir, nsim = 1000, ns = c(50, 100, 200, 400),
  mask = NULL, rs = 8, betas = rep(0, length(rs)))
}
\arguments{
\item{images}{String vector containing paths to the images used for the simulation.}

\item{data}{A dataframe with number of rows equal to the length of the images variable with corresponding covariates.}

\item{outdir}{A directory to save the output files that are used for the simulations.}

\item{nsim}{Number of simulations to setup for each sample size.}

\item{ns}{Sample sizes to evaluate for each simulation. Should be less than the number of images.}

\item{mask}{If performing simulations under an alternative, signal can be added to the images within the mask.}

\item{rs}{vector of radii for signal spheres.}

\item{betas}{vector of parameters for signal spheres. Signal is constant throughout the sphere.}
}
\value{
Returns a matrix of directories where simulation setup files are stored.
}
\description{
The images can be images from a real data set.
The bootstrap-based simulation conditions on the distribution of the sample by drawing subsets with replacement from the sample.
}
