% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genSimData.R
\name{genSimData}
\alias{genSimData}
\title{Generates data for simulation}
\usage{
genSimData(
  files,
  outfiles,
  betaimg = NULL,
  mask = NULL,
  method = c("bootstrap", "synthetic")
)
}
\arguments{
\item{files}{a vector of .nii or .nii.gz images.}

\item{outfiles}{a vector of images to save the output.}

\item{betaimg}{a parameter image that describes the association between X
and the desired relationship images in filelist. The units of this are standardized so that it
is like effect size. Currently, not supported.}

\item{mask}{image mask where data exist.}

\item{method}{Method to generate data, either "synthetic" (i.e. multivariate normal) or "bootstrap."}
}
\description{
This function loads images specified in filelist and adds signal to the
 images. The amount and structure of signal is determined by betaimg and X.
 The output images are equal the images specified by filelist plus betaimg
 times the column of X that is not in Xred. This column is first
 residualized to Xred.
}
\keyword{null}
\keyword{power}
\keyword{simulation}
