% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{runSim}
\alias{runSim}
\title{Runs simulations}
\usage{
runSim(
  simdirs,
  sims,
  simfunc,
  simfuncArgs = NULL,
  mask = NULL,
  method = c("bootstrap", "synthetic"),
  syntheticSqrt,
  ncores = parallel::detectCores(),
  ...
)
}
\arguments{
\item{simdirs}{Vector of simulation directories created by simSetup.}

\item{sims}{Vector of integers specifying simulation index. Does not have to be unique. Can be used for running a particular analysis in every kth simulation.}

\item{simfunc}{Function to evaluate on the simulated data.}

\item{simfuncArgs}{List of arguments passed to simfunc.}

\item{mask}{mask image argument passed to genSimData. Only required for genSimData function if betaimg is not NULL or if method='synthetic'}

\item{method}{method argument to generate sample data in genSimData. Data can be generated by bootstrapping or synthetically.}

\item{syntheticSqrt}{If method is synthetic this is a square root of the covariance matrix used to generate the imaging data. Passed as an RDS file or a matrix.}

\item{ncores}{Number of cores to use.}

\item{...}{Arguments passed to pbmclapply}
}
\value{
Returns the parameter image after writing it to file.path(outdir, 'signal.nii.gz').
}
\description{
Uses the mask variable to create length(rs) spheres of size rs with values equal to betas within the mask.
A random voxel is chosen within the mask and a sphere is placed at that location.
The spheres are masked by the mask image so that no parameter values exist outside the mask.
}
