% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/master.R
\name{simSetup}
\alias{simSetup}
\title{Sets up a Neuroimaging bootstrap-based simulation given a set of images and covariates.}
\usage{
simSetup(
  images,
  data,
  outdir,
  nsim = 1000,
  ns = c(50, 100, 200, 400),
  nMeas = 1,
  mask = NULL,
  ncores = parallel::detectCores()
)
}
\arguments{
\item{images}{String vector containing paths to the images used for the simulation.}

\item{data}{A dataframe with number of rows equal to the length of the images variable with corresponding covariates.}

\item{outdir}{A directory to save the output files that are used for the simulations.}

\item{nsim}{Number of simulations to setup for each sample size.}

\item{ns}{Sample sizes to evaluate for each simulation. Should be less than the number of images.}

\item{nMeas}{Approximate average number of measurements per subject.}

\item{mask}{If performing simulations under an alternative, signal can be added to the images within the mask.}

\item{ncores}{number of cores for parallel commands.}
}
\value{
Returns a data frame of directories where simulation setup files are stored. The rds file in each directory saves the data frame and the image locations in the variable "images".
}
\description{
The images can be images from a real data set.
The bootstrap-based simulation conditions on the distribution of the sample by drawing subsets with replacement from the sample.
}
